pyproject.toml
[project]
name = "journal-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "pydantic-settings>=2.2",
    "PyJWT>=2.8",
    "uvicorn>=0.27",
]
